added support for path-level parameters
According to the spec Path Item Objects can include parameters:

http://swagger.io/specification/#pathItemObject

"A list of parameters that are applicable for all the operations described
under this path. These parameters can be overridden at the operation level, but
cannot be removed there."

This commit adds support for this scenario and modifies the existing tests to
provide coverage for it.